requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0

# GitHub integration
PyGithub==2.1.1

# Task queue (Redis optional)
redis==5.0.1

# Error handling and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
flake8==6.1.0
black==23.11.0

# Production server
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1

# AI/ML libraries (for agents)
crewai==0.1.0
langchain==0.1.0
openai==1.6.1
anthropic==0.8.1

# Vector database
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
